refactor foreach
